refactor(www/audio): move UI-components into sub-folder
